Moved the data models for pydantic to new files
