+++ backend/backend/routes/cpus.py
@@ -1,11 +1,12 @@
 import os
 import re
 
-from bson import ObjectId
 from fastapi import APIRouter, HTTPException
 from motor.motor_asyncio import AsyncIOMotorClient
 
 from backend.utils.validation import validate_hardware_list
+
+from backend.models.Cpu import Cpu
 
 """
 All functions for handling the CPUs in the DB will be here for ease of use and maintainability.
